pyproject.toml
[project]
name = "marquito"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dateutil>=2.9.0.post0",
    "sqlalchemy>=2.0.48",
    "strawberry-graphql[fastapi]>=0.311.1",
    "uvicorn[standard]>=0.41.0",
]


[dependency-groups]
dev = [
    "python-semantic-release>=10.5.3",
    "debugpy>=1.8.16",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "aiosqlite>=0.22.1",
]


[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/marquito"]


[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]

build_command = """
    uv lock --upgrade-package "$PACKAGE_NAME"
    git add uv.lock
    uv build
"""

branch = "main"
changelog_file = "CHANGELOG.md"
tag_format = "v{version}"
commit_parser = "conventional"

upload_to_release = false
upload_to_pypi = false
commit_version_number = false

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf", "up"]
parse_squash_commits = true
ignore_merge_commits = true

